Filtered folders and added width n height to GameObject, and wasted time reeinventing shit Greger already had done that i didnt know of.

+++ projecttower/Pos.cpp
@@ -5,7 +5,7 @@ Pos::Pos() {
 	y = 0;
 }
 
-Pos::Pos(float xPos, float yPos) {
+Pos::Pos(double xPos, double yPos) {
 	x = xPos;
 	y = yPos;
 }
